chore(wasm-prep): kill runBlocking in commonMain (step 3)
Step 3 of the WASM pre-flight plan. wasmJs has no `runBlocking`, so two
commonMain call sites had to be replaced:

OdinId.cachedHash
-----------------
Was bridging from the synchronous `OdinId(String)` constructor (called by
the KSerializer.deserialize path, which cannot suspend) to a suspending
SHA-256 via `runBlocking`. The plan's preferred fix (precompute eagerly
in coroutine context) doesn't work because KSerializer.deserialize itself
is non-suspending — there's nowhere to suspend before construction.

Fix: introduce a synchronous SHA-256 path.

- `PureKotlinSha256.kt` ships a FIPS-180-4 SHA-256 in pure Kotlin (~90
  lines). Multiplatform, no runtime deps. Bit-for-bit equivalent to the
  existing `cryptography-kotlin` async path for all inputs we tested
  (empty, short string, block-boundary 64 bytes, multi-block 1000 bytes,
  the actual domain-name shapes that flow through OdinId).
- `HashUtil.sha256Sync(input)` exposes it; the existing async
  `HashUtil.sha256` is unchanged.
- `ByteArrayUtil.reduceSha256HashSync(...)` mirrors the async
  `reduceSha256Hash` (SHA-256 → split → XOR → 16 bytes).
- `OdinId.cachedHash` now uses the sync path. `runBlocking` removed.
- The now-unused `OdinId.calculateHash` private suspend helper is
  deleted; `OdinId.toHashId(domainName)` calls `reduceSha256Hash`
  directly.
- New `PureKotlinSha256Test` verifies the sync path matches the async
  path on five inputs and the two FIPS 180-4 test vectors.

TranslationUtil.getString
-------------------------
Was using `runBlocking { compose.resources.getString(...) }` because the
compose-resources lookup is suspending but `TranslationUtil` is used
from non-Compose code (snackbar callbacks, message rendering, etc.).

Fix: convert `TranslationUtil.getString(...)` overloads to `suspend fun`
and update the four affected call sites:

- `MessageMapper.renderStatusMessage` → `suspend`. Its only caller
  (`mapToMessageData`) was already suspending, so no further propagation
  needed.
- `AttachmentHandler.kt`: no code change — the two calls were already
  inside `scope.launch { try { ... } catch { ... } }` blocks.
- `ConversationListScreen.kt`: the three `TranslationUtil.getString`
  calls inside `onSuccess`/`onError` lambdas are now wrapped in the
  existing `scope.launch { ... }` that follows them, instead of the
  string being computed in the lambda body and then handed off.

Side-effect cleanup
-------------------
`homebase-chat/src/wasmJsMain/.../XorIdUtil.web.kt` was carrying its own
copy of the pure-Kotlin SHA-256 (added in step 2). Now delegates to
`HashUtil.sha256Sync` so we have a single source of truth.

The wasmJs target remains commented out in every module. Android / jvm /
native are functionally unchanged. JVM tests pass.
